# Avoid net.ResolveIPAddr in ParseIP