--- a/go.mod
+++ b/go.mod
@@ -1,3 +1,5 @@
 module inet.af/netaddr
 
 go 1.12
+
+require golang.org/x/perf v0.0.0-20200318175901-9c9101da8316 // indirect