--- a/netaddr_test.go
+++ b/netaddr_test.go
@@ -739,6 +739,10 @@ func TestParseIPPrefixError(t *testing.T) {
 			errstr: "no '/'",
 		},
 		{
+			prefix: "1.257.1.1/24",
+			errstr: "unable to parse IP",
+		},
+		{
 			prefix: "1.1.1.0/q",
 			errstr: "bad prefix",
 		},
@@ -786,6 +790,10 @@ func TestParseIPError(t *testing.T) {
 			ip:     "fe80::1cc0:3e8c:119f:c2e1%",
 			errstr: "missing zone",
 		},
+		{
+			ip:     "%eth0",
+			errstr: "missing IPv6 address",
+		},
 	}
 	for _, test := range tests {
 		t.Run(test.ip, func(t *testing.T) {
@@ -794,7 +802,7 @@ func TestParseIPError(t *testing.T) {
 				t.Fatal("no error")
 			}
 			if test.errstr == "" {
-				test.errstr = "unable to parse ip"
+				test.errstr = "unable to parse IP"
 			}
 			if got := err.Error(); !strings.Contains(got, test.errstr) {
 				t.Errorf("error is missing substring %q: %s", test.errstr, got)
@@ -915,5 +923,33 @@ func BenchmarkIPv4Contains(b *testing.B) {
 	ip := IPv4(192, 168, 1, 1)
 	for i := 0; i < b.N; i++ {
 		prefix.Contains(ip)
+	}
+}
+
+func BenchmarkParseIPv4(b *testing.B) {
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		ParseIP("192.168.1.1")
+	}
+}
+
+func BenchmarkParseIPv6(b *testing.B) {
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		ParseIP("fe80::1cc0:3e8c:119f:c2e1%ens18")
+	}
+}
+
+func BenchmarkStdParseIPv4(b *testing.B) {
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		net.ParseIP("192.168.1.1")
+	}
+}
+
+func BenchmarkStdParseIPv6(b *testing.B) {
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		net.ParseIP("fe80::1cc0:3e8c:119f:c2e1")
 	}
 }