ParseIP isn't documented to do dns resolution, it really shouldn't
anyway. A resolving function should be named something else, maybe even
ResolveIPAddr to match stdlib.

I originally was falling back to a combination of ParseIP for the ip and
ResolveIPAddr for the zone info, but according to RFC4007 splitting on
'%' is enough.

Signed-off-by: Manuel Mendez <mmendez534@gmail.com>

Fixes #32