@mmlb 2020-05-07
Allocs went down too:
```
benchmark                   old ns/op     new ns/op     delta
BenchmarkParseIPv4-8        255           75.5          -70.39%
BenchmarkParseIPv6-8        332           211           -36.45%
BenchmarkStdParseIPv4-8     44.1          44.0          -0.23%
BenchmarkStdParseIPv6-8     75.4          75.8          +0.53%

benchmark                   old allocs     new allocs     delta
BenchmarkParseIPv4-8        5              2              -60.00%
BenchmarkParseIPv6-8        5              3              -40.00%
BenchmarkStdParseIPv4-8     1              1              +0.00%
BenchmarkStdParseIPv6-8     1              1              +0.00%

benchmark                   old bytes     new bytes     delta
BenchmarkParseIPv4-8        132           20            -84.85%
BenchmarkParseIPv6-8        160           64            -60.00%
BenchmarkStdParseIPv4-8     16            16            +0.00%
BenchmarkStdParseIPv6-8     16            16            +0.00%
```