@mmlb 2020-05-12
Did the fixes as a fixup commit instead of fixing directly to ease review, let me know if I should squash instead.